input: disable segfaulting SDL input handling
For some ungodly reason, this code from the SDL documentation seems to
cause segfaults and XIO errors, as well (even more perplexingly) noise
in the video buffer??? I'm disabling it for now. Will have to get to the
bottom of it though, or else use something else to handle input.